Add support for hokusai apps
