+++ lib/releasecop/version.rb
@@ -1,3 +1,3 @@
 module Releasecop
-  VERSION = "0.0.4"
+  VERSION = "0.0.5"
 end
